+++ README.md
@@ -1,1 +1,4 @@
-Readme for Local Action. Will be updated soon...
+BrowserStack Actions
+
+1. Setup Test Environment: [Readme](./setup-env/README.md)
+2. Local Binary Setup: [Readme](./setup-local/README.md)
