+++ setup-env/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "local-action",
+  "name": "setup-env",
   "version": "1.0.0",
-  "description": "",
+  "description": "Setup BrowserStack Test Environment",
   "main": "src/index.js",
   "scripts": {
     "lint": "./node_modules/.bin/eslint 'src/*' 'test/**/*.js' 'config/*.js'",
     "build": "./node_modules/.bin/ncc build src/index.js",
-    "test": "nyc --reporter=html ./node_modules/.bin/mocha 'test/**/*test.js'"
+    "test": "npm run lint && nyc --reporter=html ./node_modules/.bin/mocha 'test/**/*test.js'"
   },
   "repository": {
     "type": "git",
@@ -18,16 +18,10 @@
   "bugs": {
     "url": "https://github.com/rchougule/local-action/issues"
   },
-  "homepage": "https://github.com/rchougule/local-action#readme",
+  "homepage": "https://github.com/rchougule/local-action/setup-env#readme",
   "dependencies": {
-    "@actions/artifact": "^0.3.5",
     "@actions/core": "^1.2.4",
-    "@actions/exec": "^1.0.4",
-    "@actions/github": "^4.0.0",
-    "@actions/io": "^1.0.2",
-    "@actions/tool-cache": "^1.6.0",
-    "minimist": "^1.2.5",
-    "uuid": "^8.3.0"
+    "@actions/github": "^4.0.0"
   },
   "devDependencies": {
     "@zeit/ncc": "^0.22.3",
